+++ ChangeLog.md
@@ -6,6 +6,19 @@ ___
 
 
 
+## RELEASE 3.4 - 11/03/2022
+
+- FIX : script de migration des messages llx_ticket_msg vers llx_actioncomm *17/05/2022* - 3.4.5
+- FIX : script de migration des notes de frais et compatibilité ndfp / ndfp_rh  *11/04/2022* - 3.4.4
+- FIX : Avoid emptying token value when clear filters in list *01/04/2022* - 3.4.3
+- FIX : Database integrity script update  *17/03/2022* - 3.4.2
+- FIX : Ajout d'une compatibilité pour le dépôt git ndfp et pas seulement ndfp_rh  *15/03/2022* - 3.4.1
+- NEW : Ajout du script de migration des notes de frais du module NDFP+ vers standard Dolibarr  *10/03/2022* - 3.4
+
+## RELEASE 3.3 - 14/01/2022
+
+- NEW : Script for set to 1 encrypting of the password in multicompany admin/security conf  *14/01/2022* - 3.3
+
 ## RELEASE 3.2 - 29/06/2021
 - FIX : fatal php8.1 remove &$GLOBAL to $GLOBAL - *31/05/2022)* - 3.2.9  
 - FIX : Script for MVD add missing substitutions  *08/07/2021* - 3.2.8